+++ gallery/animate-workbench/js/main.ts
@@ -22,15 +22,13 @@ $(window).on("load", () => {
     easing?: ABeamer.EasingName | ABeamer.ExprString;
     easingX?: ABeamer.EasingName | ABeamer.ExprString;
     oscillator?: ABeamer.Oscillator;
+    path?: ABeamer.Path;
   }
 
   const expr = `=iff(t<0.2 or t>0.7, 0, 1)`;
+  const pathExpr = '=[x0+(x1-x0)*t, y0-(y0-10)*t*t]';
 
   const tests: Test[] = [{
-    enabled: true,
-    label: `easing:${expr}`,
-    easing: expr,
-  }, {
     enabled: true,
     label: 'easeInQuart',
     // easing by Id
@@ -68,35 +66,64 @@ $(window).on("load", () => {
     oscillator: {
       handler: 'harmonic',
     },
+  }, {
+    enabled: true,
+    label: `path: ${pathExpr}`,
+    path: { handler: pathExpr },
+  }, , {
+    enabled: true,
+    label: `easing:${expr}`,
+    easing: expr,
   }];
 
   const h = story.height;
   const margin = 10;
   const d = 2; // duration in seconds
   const allowMultiple = true;
+  const x0 = margin;
+  const x1 = story.width - 2 * margin;
+  const y0 = h / 2;
 
   tests.forEach((test, index) => {
 
     if (index && !allowMultiple) {
       return;
     }
+
+    const props: ABeamer.AnimationProp[] = !test.path
+      ? [{
+        prop: 'left',
+        easing: test.easingX,
+        valueStart: `${margin}px`,
+        value: x1,
+      }, {
+        prop: 'top',
+        easing: test.easing,
+        oscillator: test.oscillator,
+        valueStart: `${h / 2}px`,
+        value: h / 8,
+      }]
+      : [{
+        prop: 'left-top',
+        path: test.path,
+      }];
+
     scene1
       .addAnimations([{
         selector: '#dot',
         duration: `${d}s`,
         enabled: test.enabled,
-        props: [{
-          prop: 'left',
-          easing: test.easingX,
-          valueStart: `${margin}px`,
-          value: story.width - 2 * margin,
-        }, {
-          prop: 'top',
-          easing: test.easing,
-          oscillator: test.oscillator,
-          valueStart: `${h / 2}px`,
-          value: h / 8,
+        tasks: [{
+          handler: 'add-vars',
+          params: {
+            vars: {
+              x0,
+              x1,
+              y0,
+            },
+          } as ABeamer.AddVarsTaskParams,
         }],
+        props,
       },
       {
         selector: '#label',