+++ test/tests/test-expressions.js
@@ -27,6 +27,8 @@ var Tests;
         jsMacros.push(["\"f" + i + "\"",
             "function (t) {\n          if (t === 0) {\n            var actual = ABeamer.calcExpr(\"" + expr + "\", story._args);\n          }\n      return t;\n    }"]);
     }
+    // @TODO: Implement array testing
+    // add('=[56.4, 6, 7]', 'simple array', '56.4,6,7');
     add('=  -5', 'simple value');
     add('= 4 -5', 'number sign');
     add('=  12 + 24', 'simple binary op');